+++ packages-hook/use-click-away/CHANGELOG.md
@@ -1,9 +1,14 @@
 # CHANGELOG
 
+## 0.0.3 2026/03/26 @驳是
+
+* FIX Cannot use `click` for it will hide right after open
+* FEAT Added 2nd arg `ingore`
+
 ## 0.0.2 2026/03/26 @驳是
 
-* FIX changed event from `mousedown` to `click`
+* FIX Changed event from `mousedown` to `click`
 
 ## 0.0.1 2026/03/26 @驳是
 
-* FEAT first version
+* FEAT First version